requirements.txt
py-cord
pydbantic
python-dotenv
requests
aiosqlite